Convert to mfc
