+++ resources/views/pages/inventory/⚡index/index.test.php
